requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
tqdm==4.66.1
pyyaml==6.0.1
matplotlib==3.8.2
python-dotenv==1.0.0
pyyaml
numpy
torch
scikit-learn
